requirements.txt
Django==3.2.6
psycopg2-binary==2.9.1
gunicorn==20.1.0
django-crispy-forms==1.12.0
django-tables2==2.4.0
django-extra-views==0.14.0
django-bootstrap4==3.0.1